# Switch ObjectPool to use Interlocked.Exchange instead of CompareExchange