@ToddGrun 2025-01-08
> _seems_ fine. let's run a speedometer run to ensure no unpleasant surprises :)

Test PR insertion: https://dev.azure.com/devdiv/DevDiv/_git/VS/pullrequest/601391