@ToddGrun 2025-01-08
Test insertion is being wonky during build, will work to get another one going too.

In the meantime, adding some observant eyes directly as the code is ancient and a bit scary to change. Thanks!

@AlekseyTs @jasonmalinowski @sharwell