@ToddGrun 2025-01-08
> @ToddGrun Could you clarify the goal of this PR? Fixing an problem in an algorithm, optimization, etc.

The intent is to slightly improve the efficiency of the Roslyn ObjectPool class by allowing the ObjectPool.Allocate method to return cached items that would have otherwise not been returned. By changing the code to use Interlocked.Exchange instead of Interlocked.CompareExchange, the Allocate method can find hits even when the value from the interlocked read differs from the optimistic read value.