@ToddGrun 2025-01-10
Sadly, I think I might have to abandon this PR. The speedometer tests led to some interesting results when looking at CPU/Allocations when filtering with IncPaths under PooledObjects.ObjectPool:

In the Roslyn code analysis process (.net core), things look good:
  Allocations went from 36MB -> 29.7 MB
  CPU went from 939 ms -> 689 ms

However, in the devenv process (netfx), things are mixed: 
  Allocations went from 75.6 MB -> 50 MB
  CPU went from 567 ms -> 1384 ms.

Two different speedometer runs had the same trends, things improved except for CPU in the VS process. @stephentoub -- sorry to pull in the big guns here, but is it to be expected that Interlocked.CompareExchange and Interlocked.Exchange would have such different performance characteristics in these two contexts?