@ToddGrun 2025-01-13
Closing as this didn't degraded CPU perf in VS as outlined above. Seems a bit counter-intuitive, but it's what the tests demonstrate.