@stephentoub 2025-01-17
On the surface I'm not sure why you'd see such a significant CPU difference between core and netfx here; there have been improvements around {Compare}Exchange, but I wouldn't expect them to be so significant. My guess is there's something else subtle going on, e.g. Exchange is always going to incur a write whereas CompareExchange won't, so maybe there's something about the memory layout that's contributing to a false sharing issue on netfx, or something like that. Hard to say without spending more time measuring and analyzing.